+++ src/assets/bars/src/scripts/types/Craft.d.ts
@@ -3,5 +3,6 @@
  */
 declare const Craft: {
   MatrixInput: GarnishComponent
+  Smith: any
   t: (category: string, message: string, params?: object) => string
 }